requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
defusedxml==0.7.1
dnspython==2.4.2
fastapi==0.104.1
h11==0.14.0
idna==3.4
jwt==1.3.0
multidict==6.0.4
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.6
uvicorn==0.24.0
yarl==1.9.2
